listening-comp/backend/requirements.txt
fastapi
uvicorn
youtube_transcript_api
google-genai
soundfile
python-dotenv
boto3
